+++ projects/ngx-state-traverser/src/lib/reducer.ts
@@ -1,4 +1,4 @@
-import { initialState, TraversingState } from './state';
+import { initialState, TraverserState } from './state';
 import { TraverserActions } from './actions';
 
 /**
@@ -31,7 +31,7 @@ export function deepMerge(target: any, source: any) {
     return target;
 }
 
-export function reducer(state = initialState, action: TraverserActions.Actions): TraversingState {
+export function reducer(state = initialState, action: TraverserActions.Actions): TraverserState {
     switch (action.type) {
         case TraverserActions.Types.ResolveContext: {
             let contextPath = action.payload.contextPath;
@@ -107,6 +107,27 @@ export function reducer(state = initialState, action: TraverserActions.Actions):
                 return {
                     ...state,
                     target: {...state.target, context: resource},
+                    collection,
+                };
+            } else {
+                return { ...state, collection, };
+            }
+        }
+        case TraverserActions.Types.AddOrUpdateTraverserResources: {
+            let newContext: any;
+            const collection = action.payload.reduce((acc, value) => {
+                const updatedObject = !!state.collection[value.path] ? deepMerge({...state.collection[value.path]}, value.changes) :
+                    value.changes;
+                if (state.target.contextPath === value.path) {
+                    newContext = updatedObject;
+                }
+                acc[value.path] = updatedObject;
+                return acc;
+            }, {...state.collection});
+            if (!!newContext) {
+                return {
+                    ...state,
+                    target: {...state.target, context: newContext},
                     collection,
                 };
             } else {