+++ projects/ngx-state-traverser/src/lib/store.ts
@@ -1,8 +1,8 @@
 import { StoreModule, StoreFeatureModule } from '@ngrx/store';
 import { ModuleWithProviders } from '@angular/core';
 import { reducer } from './reducer';
-import { TraversingStateFeatures } from './state';
+import { TraverserStateFeatures } from './state';
 
 export const store: ModuleWithProviders<StoreFeatureModule>[] = [
-    StoreModule.forFeature(TraversingStateFeatures.Traversal, reducer)
+    StoreModule.forFeature(TraverserStateFeatures.Traversal, reducer)
 ];