+++ projects/nst-test/src/app/navigation/navigation.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { GithubFile } from '../file/file.component';
 import { Store } from '@ngrx/store';
-import { TraverserSelectors, TraversingState } from 'ngx-state-traverser';
+import { TraverserSelectors, TraverserState } from 'ngx-state-traverser';
 
 @Component({
   selector: 'app-navigation',
@@ -11,6 +11,6 @@ export class NavigationComponent {
 
   parent = TraverserSelectors.TraverseToParent<GithubFile>(this.store);
 
-  constructor(private readonly store: Store<TraversingState>) { }
+  constructor(private readonly store: Store<TraverserState>) { }
 
 }